refactor other JS files & move into index
