--- scripts/emailForm.js
@@ -1,1 +1,0 @@
-$('.getUpdates').click(() => $('.emailForm--inputBox').focus())
