--- scripts/header.js
@@ -1,9 +1,0 @@
-function toggleNavModal() {
-    $('#navModal').toggleClass('navModal-is-shown')
-    $('body').toggleClass('l-noScroll')
-}
-
-$('#hamburger')
-    .add('#closeNavModal')
-    .add('.navModal--linkList > *')
-    .click(toggleNavModal)
