--- scripts/preload.js
@@ -1,1 +1,0 @@
-$(document).ready(() => $('body').removeClass("u-preload"))
