--- scripts/testimonials.js
@@ -1,9 +1,0 @@
-function cloneCarousel(el) {
-    const clone = $(el).clone()
-        .removeClass('u-onlyDesktop')
-        .removeClass('carousel-double')
-        .addClass('u-onlyMobile')
-    $(el).after(clone)
-}
-
-$('#testimonials .carousel').each((idx, el) => cloneCarousel(el))
